Fixed minor issue

+++ src/backend/db_queries/announcements/deleteOffer.js
@@ -13,27 +13,30 @@ function deleteOffer(data) {
     const [id_author, author, id_offer, candidates, passengers] = data.parameters;
 
     // Removes all applications and sends a notification to each candidate
-    candidates.map((value) => {
-        if (value) { 
-            let id_candidate = parseInt(value.split(':')[0]);
-            queries.push(`INSERT INTO notification VALUES (DEFAULT, ?, CONCAT( ? , ' a annulé son offre'), false, NOW());`)
-            parameters.push([id_candidate, author])
-            queries.push(`DELETE FROM apply_offer WHERE candidate = ? AND id_offer = ? and author = ?`)
-            parameters.push([id_candidate, id_offer, id_author])
-        }   
-    })
+    if (candidates) {
+        candidates.map((value) => {
+            if (value) { 
+                let id_candidate = parseInt(value.split(':')[0]);
+                queries.push(`INSERT INTO notification VALUES (DEFAULT, ?, CONCAT( ? , ' a annulé son offre'), false, NOW());`)
+                parameters.push([id_candidate, author])
+                queries.push(`DELETE FROM apply_offer WHERE candidate = ? AND id_offer = ? and author = ?`)
+                parameters.push([id_candidate, id_offer, id_author])
+            }   
+        })
+    }
     
     // Removes all passengers and sends them a notification each
-    passengers.map((value) => {
-        if (value) { 
-            let id_passenger = parseInt(value.split(':')[0]);
-            queries.push(`INSERT INTO notification VALUES (DEFAULT, ?, CONCAT( ? , ' a annulé son offre'), false, NOW());`)
-            parameters.push([id_passenger, author])
-            queries.push(`DELETE FROM offer_client WHERE client = ? AND id_offer = ?`)
-            parameters.push([id_passenger, id_offer])
-        }   
-    })
-
+    if (passengers) {
+        passengers.map((value) => {
+            if (value) { 
+                let id_passenger = parseInt(value.split(':')[0]);
+                queries.push(`INSERT INTO notification VALUES (DEFAULT, ?, CONCAT( ? , ' a annulé son offre'), false, NOW());`)
+                parameters.push([id_passenger, author])
+                queries.push(`DELETE FROM offer_client WHERE client = ? AND id_offer = ?`)
+                parameters.push([id_passenger, id_offer])
+            }   
+        })
+    }
     // Finally delete offer from database
     queries.push(`DELETE FROM offer WHERE id = ?`)
     parameters.push([id_offer]);
